+++ package.json
@@ -11,7 +11,7 @@
     "react-dom": "^16.7.0",
     "react-redux": "^6.0.0",
     "react-router-dom": "^4.3.1",
-    "react-scripts": "2.1.2",
+    "react-scripts": "^2.1.2",
     "redux": "^4.0.1",
     "redux-devtools-extension": "^2.13.7",
     "redux-thunk": "^2.3.0"
@@ -38,7 +38,7 @@
       "src/**/*.{js,jsx}",
       "!<rootDir>/node_modules/",
       "!<rootDir>/src/themes/*",
-      "!<rootDir>/src/ProtectedRoute/*",
+      "!<rootDir>/src/privateRoute/*",
       "!<rootDir>/src/index.js"
     ]
   },
@@ -54,6 +54,10 @@
     "eslint-plugin-prettier": "^3.0.0",
     "eslint-plugin-react": "^7.11.1",
     "jest": "^23.6.0",
-    "redux-devtools-extension": "^2.13.5"
+    "react-test-renderer": "^16.7.0",
+    "redux-devtools-extension": "^2.13.5",
+    "babel-jest": "^23.6.0",
+    "babel-preset-env": "^1.7.0",
+    "babel-preset-react": "^6.24.1"
   }
 }
