+++ src/index.js
@@ -2,16 +2,16 @@ import ReactDOM from 'react-dom';
 import './static/styles/index.scss';
 import React from 'react';
 import { Provider } from 'react-redux';
+import MuiThemeProvider from '@material-ui/core/styles/MuiThemeProvider';
 import store from './store';
-import MuiThemeProvider from "@material-ui/core/styles/MuiThemeProvider";
-import AppTheme from "./static/themes";
-import Routes from "./routes";
+import AppTheme from './static/themes';
+import Routes from './routes';
 
 ReactDOM.render(
-    <MuiThemeProvider theme={AppTheme}>
-        <Provider store={store}>
-            <Routes />
-        </Provider>
-    </MuiThemeProvider>,
-    document.getElementById('root')
+	<MuiThemeProvider theme={AppTheme}>
+		<Provider store={store}>
+			<Routes />
+		</Provider>
+	</MuiThemeProvider>,
+	document.getElementById('root') || document.createElement('div')
 );
