+++ src/store/index.js
@@ -1,7 +1,7 @@
 import { applyMiddleware, createStore } from 'redux';
 import thunk from 'redux-thunk';
 import { composeWithDevTools } from 'redux-devtools-extension';
-import reducers from './reducers';
+import index from '../reducers';
 
-const store = createStore(reducers, composeWithDevTools(applyMiddleware(thunk)));
+const store = createStore(index, composeWithDevTools(applyMiddleware(thunk)));
 export default store;
